LU-9859 libcfs: add support for Xarray
In the late 4.X kernel cycle Xarrays were introduced with the
goal of replacing the radix tree for the page cache. It is
highly optimized for densely packed data which is the case for
several items in Lustre such as the static array for the obd
devices and quota ids. This patch provides Xarray support for
kernels that lack. The current verison of Xarray back ported
is from the 5.4-rc2 kernel.

Change-Id: I54f9046f50a353e1cd4271c0b97207062bbf3898
Signed-off-by: James Simmons <jsimmons@infradead.org>
Reviewed-on: https://review.whamcloud.com/37391
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Neil Brown <neilb@suse.de>
Reviewed-by: Oleg Drokin <green@whamcloud.com>