LU-13187 tests: improve sanity test_129 checking
When running sanity test_129 repeatedly, it may fail to print the
warning/error messages to the console due to throttling.  Fix the
test to only warn if the messages were not printed, so long as the
size limit is correct.

Always reset max_dir_limit on exit, and delete the test files.

Fix test_129 to use mcreate instead of multiop to be faster, and
fix mcreate to return the actual error code returned by the syscall.

Fix the console error messages to be complete strings, rather than
being dynamically generated, so that they can be more easily found.
To compensate, don't have separate messages for with/without FID.

Test-Parameters: trivial testlist=sanity env=ONLY=129,ONLY_REPEAT=100
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I06593441151fa508ef2db7a3e68be34cc0ce7057
Reviewed-on: https://review.whamcloud.com/38550
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Li Dongyang <dongyangli@ddn.com>
Reviewed-by: James Nunez <jnunez@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>