LU-13151 mdt: add parent FID to Changelog recordss
Use the link EA to add the parent FID to ChangeLog records, including
MTIME, TRUNC, and SATTR.

Some tools that maintain copies of filesystem metadata in an external
database monitor changelogs for changes to the filesystem, in order to
determine files that need to be rescanned. This can result in a large
number of small updates to the external database that can reduce the
tool's ingest performance. It might be beneficial to instead track and
scan complete directories that contain modified files and update the
external database using bulk operations. Adding the parent FID to
MTIME changelogs allows to more efficiently determine the parent
directories for some types of file data modifications, by issuing
OBD_IOC_FID2PATH once for each parent FID, instead of once for each
file FID.

Signed-off-by: Nikitas Angelinas <nikitas.angelinas@hpe.com>
Cray-bug-id: LUS-7986
Change-Id: I0c88271e706ef8202910e9461e5ae9f6dcbe0bdd
Reviewed-on: https://review.whamcloud.com/37264
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Nathan Rutman <nrutman@gmail.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>