LUDOC-11 misc: remove references for older Lustre
Remove references to Lustre 2.4 and earlier from the manual.
There were even places with referencs to Lustre 1.8 and 2.0.

The text still exists in earlier builds and in Git for reference.
This avoids confusing users with commands that no longer exist,
or highlighting features that are included with all releases.

Add conditions for Lustre 2.15 and 2.16 as they will be used soon.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I0786d88fcc2543f542b7a880bf042928fd3ebbe5
Reviewed-on: https://review.whamcloud.com/39511
Tested-by: jenkins <devops@whamcloud.com>