LU-13773 tests: use TESTLOG_PREFIX in run_one_logged
TESTLOG_PREFIX is defined and exported in init_test_env()
in test-framework.sh.  This environment variable is defined
as $LOGDIR/$TESTSUITE. TESTLOG_PREFIX should  be used in the
definitions of test_log and zfs_debug_log.  Since the logs
created in run_one_logged() don't use the defined prefix,
this could lead to differences in the naming of the dmesg,
debug_log and test_log logs.

Let's use TESTLOG_PREFIX in the definitions of test_log and
zfs_debug_log so that changes to the prefix are reflected in
all logs.

Signed-off-by: James Nunez <jnunez@whamcloud.com>
Change-Id: I14dbe0469b7a6627d63679103c235b97f0e42b67
Reviewed-on: https://review.whamcloud.com/39552
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Wei Liu <sarah@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Jian Yu <yujian@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>