LU-11310 ldiskfs: Fix suse15/ext4-max-dir-size.patch
The ext4-max-dir-size patch for suse15 added a 'max_dir_size' sysfs
attribute with an incorrect implementation.  The implementation is
identical to that for 'max_dir_size_kb', so setting or reading
'max_dir_size' will result in incorrect values.  This causes
sanity test 129 to fail.

So add a suitable implementation for max_dir_size

Test-Parameters: trivial
Signed-off-by: Mr NeilBrown <neilb@suse.de>
Change-Id: I591259ed668bc828c3a7caa6a55d0de2b0d72797
Reviewed-on: https://review.whamcloud.com/39571
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Petros Koutoupis <petros.koutoupis@hpe.com>
Reviewed-by: Jian Yu <yujian@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>