LU-13902 config: add test for /usr/include/libiberty/
The Makefile currently assumes that -I/usr/include/libiberty
is only needed on UBUNTU.  But UBUNTU is only evaludate if
modules are being built (it is based on the base kernel found).
So with --disable-modules, -I/usr/include/liberty is never
selected.

There is no need for this indirection.  It is easy to check if
/usr/inclue/libiberty exists, and to use if it if does.
So do that.

Signed-off-by: Mr NeilBrown <neilb@suse.de>
Change-Id: I6fb482256df3c7ce537dd4d451a9da27b8d8d3ea
Reviewed-on: https://review.whamcloud.com/39749
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>