LU-13019 tests: replace usleep in test scripts
The "usleep" command is specific to RHEL, and is not available on
Ubuntu clients.  Instead, "sleep" can be used with a fractional
second value to get the same effect more portably.

The "usleep" values are increased to be more reasonable lengths, as
a 100us sleep doesn't make sense after do_facet or other remote call
that may take seconds to complete.  Also, in VM testing there may
be unpredictable delays, so it is better to be conservative.

Test-Parameters: trivial testlist=sanity-hsm,replay-single,replay-dual
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: Ic88c37d33ac17f46b9c1fb912112d318d9c3d07f
Reviewed-on: https://review.whamcloud.com/39889
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Emoly Liu <emoly@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Reviewed-by: James Nunez <jnunez@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>